pmrep: improve corner case dupe detection
I haven't been able to reproduce the issue fixed in a meaningful way
to create a QA check but it basically always involved all the sample
metrics, including those with dupe names, and some instance filtering
and non-sample metrics, and a bit of luck.  But looking at the change
it makes sense and the current QA is still happy.