src/pcp/atop/various.c: fix logic error in "forever" time calculation
In the "run forever" case (as exercised in qa/1089), the number of
samples is UINT_MAX, multiply this by the default sample interval
(10) and you get a "duration" that when forced into a tv_sec may be
negative ...  causing atop to exit "normally" immediately.

Avoid arithmetic overflow by cropping the "duration" to INT_MAX-1.