build: debian python packaging fixups
There is a separate depends variable for python3, python3:Depends.
Using the wrong one for python3-pcp was causing the python3 interpreter
related depends to be missing.  It also removes a few obsolete items -
the variable python3:Provides has never existed and python:Provides is
not to be used any longer.

Resolves Debian bug #866520.