libpcp_pmda: small changes, mostly additional diagnostics
When -Dlibpmda,desperate is used, there is some low-level tracing
of calls into libpcp_pmda ... only pmdaDesc() and pmdaFetch() are
done at this stage, as that was sufficient to debug a nasty problem
with the news PMDA on OpenBSD (vm33).

Also added a paranoid guard before the callback to
(pmda->e_resultCallBack)(). This was also related to the
OpenBSD debugging, but this one was not implicated in the end
because pmda->e_resultCallBack is set to __pmFreeResultValues()
via pmdaSetResultCallBack() in __pmdaSetup() and is unlikely to be
changed, much less set to NULL.  Kept the guard as it is benign and
matches the style of the other callback code.