src/pmdas/news/pmdanews.pl: fix ps(1) usage
Buried in the Perl code is a ps(1) command ... but the syntax was
of the SysV/Linux style, not the BSDian style, so it failed miserably
on OpenBSD (vm33).

We already have $PCP_PS_PROG and $PCP_PS_ALL_FLAGS in /etc/pcp.conf
for _just_ this reason, so get the values from there via a couple of
pmda_config() calls.