pcp-mpstat: gracefully handle the absence of interrupt metrics
Noticed that pcp-mpstat fails to run when per-CPU soft and hard
interrupt metrics are missing from the source (archive, although
it will probably affect non x86 hosts as well I suspect).

For the default report case these are not actually used, so this
is quite unfortunate - the tool can still produce correct output
in that case (now does).

Added an archive showing the problem and extended qa/883.