pmdadm: small changes from review of dmstats code
Rearrange cpp logic so that all conditional code is now
in the dmstats.h header to reduce the spread of #ifdef
throughout the PMDA.  Drop an empty setup function and
a commented-out structure definition.  Dropped use of
the unconditionally-set HAVE_DMSTATS macro, and drive
this logic from makefile CFLAGS (-DHAVE_DEVMAPPER).

Revert an accidental (I think?) change to the "cache"
target name, was correct previously.