labels: add support for a cluster level of label hierarchy
After looking at the needs of MMV and Prometheus PMDAs, we
observe a need to set labels at the PMID cluster level (so
for labels common to individual instrumented applications,
IOW).  This commit splits PMID labels into cluster/item.

Some smaller API improvements are made while revisiting the
PMDA interface at this time.  Helper routines are added to
allow easy creation of notes (optional labels) and also to
set flags on a labelset.  Additionally, the pmdaLabel code
is updated to be more like the other helpers (pmdaFetch, et
al) in that it should always be called by the PMDA label
routine as it provides default handling of labelsets at all
levels now.  This allows the correct flags to be provided
in every labelset returned.

The pmGetLabels routine is now separated from the instances
routine, and semantics changed to provide the fully merged
label set for a single metric.