libpcp: small derived metrics metadata change
For expressions involving division or multiplication, the semantics
of the result may involve the dimension (for space, time or count)
becoming zero.  This part was all OK, but when this happens the
scale should also be set to zero (rather than inherited from one
or both of the operands) ... most cases this does not matter because
once the dimension is zero the scale is not inspected, but the one
exception is for ternary expressions where the left and right
expressions must have identical metadata.

Trying to fix Marko's "IO percentage for a process" expression
exposed just this problem!

Extended qa/938 to cover this.

In addition to this change, I added some diagnostics to help debug
the problem and in the process made the use of the debug options
(-Dderive with some/all of -Dappl0,appl1,appl2) consistent, which
triggered a couple of QA fallouts.