Fix use after free in perf_setup_dynamic_events().
There is a use after free bug in the perf_setup_dynamic_events(). Steps
to reproduce:

1) Run the perfevent PMDA using the default configuration file on a host
that does not have a /sys/bus/event_source/devices/*/cpumask file (i.e. Kernel version < 3.7).
2) Observe that it almost always segfaults.

I've added an extra test for this scenario into the unit test harness.
Ran with valgrind to confirm that no new memory leaks are created by
this change. (I note that there are a few existing 'leaks' in the test
harness. I assume most are due to the lack of cleanup code in the test
harness itself.)