labels: indom identifier, not pmid, for instance labels
Switch to passing the instance domain identifier as the
key for PM_LABEL_INSTANCES, not the PMID.  This is more
correct in terms of identifying the entity the label is
associated with, and the intent is now much more clear.

Some fallout remains for the Prometheus endpoint PMDA;
I'll followup with Ronak on that one and we'll resolve
it in a subsequent commit.