src/libpcp/src/derive_fetch.c: fix ternary expression evaulation bug
This was the second part of Marko's iopct bug.

If a ternary operator was used, and the true/false expressions were
singular and over an indom (not the otherway round) and the indom
side had exactly one value in the pmResult and the guard expression
was false ... then we returned the correct value, just for instance
PM_ID_NULL instead of the instance from the false indom expression.

Needless to say, this is a bit of a corner case, which explains why
it had not been seen before!

qa/1227 now covers all the permutations here for the guard's
truth/false value and the true/false operands being singular and
associated with an indom or both being associated with (the same)
indom.