Fix __pmFILe compilation failures on *BSD
Changed all the method names in the __pm_fops struct to avoid possible
compiler name clashes with different stdio implementations, e.g.
change from
    int (*fileno) (__pmFILE *);
to
    int (*__pmfileno) (__pmFILE *);

Note the struct elements are all lower case.  This is not to be
confused (says he hopefully) with the abstraction wrapper methods
that are camel case, e.g. __pmFileno().

Also include <sys/stat.h> so "struct stat *" can be used in
prototypes in impl.h.