src/pmcd/pmdaproc.sh: better checks on domain number and name matching
Following a problem reported by Henry Chang and Nathan Scott,
the PMDA Install/Remove infrastructure has been updated to enforce
the (previously implicit) rule that _both_ the domain name and
the domain number must be unique in pmcd.conf.

If an attempt is made to add or delete an entry where one or other
of the domain name or domain number is not as expected, then a
warning will be issued.