qa: add qa/1096 for mutithreaded pmNewContext/pmDestroyContext crash
pmmgr/libpcp crashed when one thread was creating a new host context,
and another thread was in error handling after failing to create an
archive context on an empty archive (empty because /var/log was full).
These two threads race, leading to either (a) a use-after-free corruption
of ctxp->c_archctl that should not have been reused because it had been
freed by the thread that was in error handling, or (b) a double free
of same in pmDestroyContext.

Similar issue occurs in error handling if the archive thread fails in
other ways, e.g. an archive base name that doesn't exist, etc.