pmlogconf: cull missing groups with -c
When -c is used with an existing configfile then we're in auto-create
mode, so the right thing to do when a group is not installed is to cull
the group from the configfile.

And sent warning messages to stderr to separate them from the more
interactive chatter that happens on stdout ... makes it easier for
QA also.

Also clarified the man page in respect of this new behaviour and
added some verbage around the -c explanation.

This addresses the issue of the zeroconf/* groups that escaped and
then were repackaged on some platforms so they were no longer installed
below $PCP_VAR_DIR/config/pmlogconf.