qa/348: update list of newer kvm metrics
vm12 is running Fedora 26 and with this comes a new Linux kernel
with new kvm metrics we've not seen before, kvm.req_event and
kvm.max_mmu_page_hash_collisions.

Add these to the ones we check but don't report a positive
(in range) outcome for, so 348.out is version-independent.