pmrep: add support for new derived metrics syntax
Adding the new iopct formula to the pmrep.conf (as opposed to a derived
metrics configuration file) revealed that pmrep didn't handle the new
derived metrics syntax which uses comma which is also used by pmrep as a
separator. Add support for this, on command line these kind of
expressions are not supported as comma was already reserved for
separating derived metric expression. But looking at the iopct example
it seems quite unlikely anyone would want to construct such a complex
expression on the command line so I think it's ok.