python api: allow long options without a short option
This is now done and similar to the way getopt(3) works;
the option callbacks final parameter is an index (always
was) indicating the offset into the long option array.

The pmSetLongOption python API now returns this index so
that the caller can distinguish individual long options.

Fixes https://github.com/performancecopilot/pcp/issues/334