python api: annotate the pmsubsys module as "deprecated"
Since pmsubsys is not a general-purpose module, doesn't really
make use of modern PMAPI services and contains x86_64-specific
interrupt metric handling with no fix planned, it's now marked
as "deprecated" using the python warnings module.  It is being
removed from the pcp-4 branch.

The only remaining user of the pmsubsys module was pcp-collectl
so the code has been shifted there.  This tool is known to be
in need of rewrite (before extension); if anyone is interested
in taking that work item on, please do!