libpcp: avoid valgrind warning for uninitialized value in __pmLogChkLabel
The following valgrind warning occurs running qa/src/archctl_segfault,
which attempts to open an archive context for qa/archives/empty :

==8001== Command: src/archctl_segfault
==8001==
==8001== Conditional jump or move depends on uninitialised value(s)
==8001==    at 0x4E6CAEA: __pmLogChkLabel (logutil.c:204)
==8001==    by 0x4E6D0AF: __pmLogChangeVol.part.2 (logutil.c:561)
==8001==    by 0x4E6E51F: __pmLogOpen (logutil.c:1158)
==8001==    by 0x4E4F369: __pmFindOrOpenArchive (context.c:583)
==8001==    by 0x4E510FF: initarchive (context.c:788)
==8001==    by 0x4E510FF: pmNewContext (context.c:1112)
==8001==    by 0x400644: main (archctl_segfault.c:19)
==8001==  Uninitialised value was created by a stack allocation
==8001==    at 0x4E4C7B8: ??? (in /usr/lib64/libpcp.so.3)

This is caused by 'len' being uninitialized on the stack :

    171 __pmLogChkLabel(__pmLogCtl *lcp, __pmFILE *f, __pmLogLabel *lp, int vol)
    172 {
    173     int         len;
    174     int         version = UNKNOWN_VERSION;
    175     int         xpectlen = sizeof(__pmLogLabel) + 2 * sizeof(len);
    176     int         n;
...
    201     __pmFseek(f, (long)0, SEEK_SET);
    202     n = (int)__pmFread(&len, 1, sizeof(len), f);
    203     len = ntohl(len);
--> 204     if (n != sizeof(len) || len != xpectlen) {
    205         if (__pmFeof(f)) {
    206             __pmClearerr(f);
...

On big-endian, ntohl does nothing and if the __pmFread returns 0 or an error
then 'len' will not be modified. So the conditional at line 204 may depend on
the uninitialized value of the len stack variable. In practice, n would be
zero and so n != sizeof(len) would always be true, so I don't think this can
happen anyway!