libpcp + pcpintro(1): add $PCP_IGNORE_MARK_RECORDS
This new environment variable can be used to clobber the
default handling of <mark> records in archives, especially when
the data is likely to be continuous in the region of the <mark>
records, e.g. after log rotation and subsequent concatenation
with pmlogextract(1).

This is a pretty crude big hammer, but is the only option for existing
archives.

A newer, smarter, protocol for pmlogger and pmlogextract will come
as part two of the fixes in this area, and this will make
$PCP_IGNORE_MARK_RECORDS redundant for archives created with the
new protocol.