libpcp: pmsprintf - a safer [v]snprintf wrapper
Follow up on one of the Canonical security team
members suggestions and introduce a wrapper for
vsnprintf that can be used more safely.

This provides guarantees that the string buffer
is null terminated (even when truncated, or in
the presence of failure in the underlying call).

Begin the process of introducing it into several
libraries.  It'll become more pervasive in time.