src/pmdas/linux/mk.rewrite: fix significant typo
The kernel.release.uname metric was incorrectly specified as
kernel_uname.release which meant the real metric was never rewritten.

This has been broken since late May 2017, but I only noticed it when
I downgraded to an older PCP release for some unrelated testing
and ended up with some archives containing the old semantics for
kernel.release.uname and this caused my cron_daily processing to fail
from then on ... it would have stopped failing when the cull period
for archives had been reached, but I noticed this first.