qa/035: rework
It turns out that the only real way to make pmrep fail
reliably is to try and process the _whole_ archive.

And with the guards added in libpcp around the derived
metric expression evalation (amongst other places) in the
previous commit, this test now consistently produces the
correct results ... a bunch of output followed by
    pmrep: Corrupted record in a PCP archive log
when we hit the corruption in the
badarchives/archive-20141125-000009 archive.