libpcp et al: change PCP debugging control infrastructure
We have historically used bit-fields over the int pmDebug and the
DBG_TRACE_xxx macros to support the -D and __pmParseDebug() services
for run-time conditional debugging and tracing in various PCP
libraries and applications.

We've nearly run out of bits in the 31-bit int (pmDebug has to stay
positive) and although pmDebug2 was suggested (to follow the XFS
thinking), there is a better way ... an extensible vector of ints
(one per debug option) that can efficiently be set and tested.

This code has (mostly) been wrapped in #ifdef PCP_DEBUG ... but the last
time this was NOT built into the code was performance tests in early
1994 to convince Akmal Khan that PCP could compete with sar/vmstat in
terms of sysem overhead to collect performance data.  I think that use
case has long since become nothing more than a quaint historical anomaly.

So this reimplementation:
- removes the restriction on at most 31 debug options or classes
- is backwards compatible, so all uses and abuses of pmDebug are
  maintained for the bit-field based debug options
- exposes a sane pmSetDebug/pmClearDebug interface as a full PMAPI
  member
- starts to peel away the #ifdef PCP_DEBUG guards