src/libpcp/src: major change to derived metrics implementation
This has been a long-standing personal itch, where the growing set
of default derived metrics imposes a growing start-up overhead on
every PCP application ... the problem is in the way I chose to do
the binding and semantic checks in the derived metric implementation.

The change moves the timing of derived metric semantic binding from
occurring for _all_ metrics immediately after each pmNewContext()
call, to lazy binding only when a derived metric's value or the derived
metric's metadata are requested for the first time in each context.

So if a PCP application does not reference any derived metrics then
no semantic binding is done.

Syntactic checking of derived metrics still happens once at each
application startup, so that derived merics appear in the PMNS.
But this is a comparatively light-weight process.

With the current default derived metrics setup, this changes the
number of PDU round-trips for

    $ pminfo -f disk.dev.total

from 115 to 7!