qa updates for after change to derived metrics implementation
There are a lot of .out files that had to be remade and audited.

If the qa test uses -Dderive for an application, then the diagnostic
output is now quite different because the semantic binding may not
be done at all for unreferenced derived metrics, or may be done at
a different point in time.