pmrep: handle dup names
It turned out that pmrep didn't handle dup names properly.

The fix is trivial but the else branch following it is a mystery,
removing it would have no QA effect but there must have been some reason
I've added it (might have been a corrupted archive or such) so leaving
it there for now.

Add a basic string output test (which were few) while at it.