libpcp: initialise unaligned part of help text buffer
Use the same approach taken in the namespace PDU code,
and initialise the trailing end of help text PDUs when
they are unaligned to __pmPDU size.

pducheck happened to be using an aligned text buffer,
so valgrind use in qa/386 didn't see this - extend it
by a byte and it exercises the fix.  qa/remake 367.

Fixes https://github.com/performancecopilot/pcp/issues/340