libpcp and pmdbg.h: restructure
pmdbg.h has the debugging options support and it is generated from
impl.h.

The details in this header are really internals for libpcp, and as
such should not be exposed.

This commit moves pmdbg.h (and the script that generates it)
src/include/pcp to src/libpcp/src.

I've retained an empty stub in src/include/pmdbg.h and we'll continue
to install this for backwards compatibility (for code that may have
included pmdbg.h, even though there was no use or value in doing that).