build: enable Qt5 builds for non-rpm platforms
After several rounds of patch ping-pong between Nathan and Ken, this
is the first cut of changes to enable building with Qt5 for the
non-rpm builds.  When both Qt4 and Qt5 are installed, we will build
with Qt5.

All of this is triggered by the intention of retiring Qt4 in the
Debian Buster (aka Debian 10) release.