pcp-atop: add --hotproc option for using hotproc metrics
Provide a pcp-atop(1) --hotproc command line option allowing
the hotproc metrics to be used instead of proc.  Example log
with hotproc metrics added for QA, several tests updated and
docs added to the man page.  Additionally, those long-option
PCP options which are directly available in pcp-atop are now
documented as such, in the man page and the usage message.

While testing, found & fixed a case where pmlogconf attempts
to cat a non-existent file - handled it the same way it does
other files in this situation.  Also made a pcp-atop archive
(used for QA) compressed.