src/libpcp/src/interp.c: harden up against bad archives
qa/317 and Marko's pmrep experiments have exposed a problem when the
encoding of a pmResult in an archive uses the wrong valfmt compared
to the expected value based on the metric's type.

Add an early test when this condition is encountered and return
PM_ERR_LOGREC ... most well-behaved applications will throw in the
towel when pmFetch() returns PM_ERR_LOGREC.

Apps that don't use interpolation mode will miss this test, but they
may not have handy access to the metric's pmDesc (and hence type) at
the point where the archive is read ...  but most of these will end
calling pmExtractValue() at some point later (when the type is known)
and a semantically similar test was already in place.

This explains why pmlogsummary and pmlogcheck were previously OK
(they don't use interpolate mode, but do call pmExtractValue).
And why pmrep (and pmie and pmval) all SEGV'd because they use
interpolate mode and died deep in the pmFetch interpolation before
getting a chance to call pmExtractValue().