python api: fix timezone segfault from incorrect free
Reading pmWhichZone(3) and src/libpcp/src/tz.c suggests that the
returned value should not be free()'d. Also, the cryptic "Operation
not permitted" is explained by the fact that initially the tzhandle
is -1 in tz.c when no timezone has been set thus the error is mapped
to EPERM on Linux.