python api: fixes to timezone handling
The patch below should finally fix all the timezone related corner cases
on Python level, making things correct and  consistent for users across
the globe regardless of their respective timezones.

The explain the situation a bit, here's what's going on:

1) PCP on C level (=libpcp) uses the POSIX-style timezone specification
(http://pubs.opengroup.org/onlinepubs/9699919799/functions/tzset.html,
http://man7.org/linux/man-pages/man3/tzset.3.html describe this) to be
used with the TZ environment variable where, unintuitively for most
people I would say, timezones east of the Prime Meridian, i.e., where
time is ahead of UTC, have negative offset (so we see something like
EST+5 for the US and AEST-11 for Australia).

2) In addition being unintuitive, timezones in POSIX/TZ-style may be
misleading for the uninitiated ("Does AMT+4 refer to Amazon or Armenia
as the former is 4 hours behind the latter is 4 hours ahead of UTC?") or
ambiguous ("What is the offset in :Europe/Helsinki and does it refer to
Standard Time or Daylight Saving time?").

3) At least one older PCP component (sheet2pcp(1)) is using non-POSIX
timezone specification where timezones west of the Prime Meridian have
positive offset (as ISO 8601 does).

4) Higher level languages like Python tend to use ISO 8601 UTC offsets
when exposing them to users and some of the recently added pcp2* Python
tools also follow this pattern.

5) In any case, with PCP tools the end result should be clear, without
the -z/-Z options the system timezone should be used, with -z the source
archive/host timezone should be used, and with -Z the given timezone (in
POSIX/TZ format) should be in effect. (This is of course a bit confusing
with some of the Python tools where with -Z UTC+3 one would actually see
UTC-3 being the timezone in use, we should actually add a sentence or
explaining this more clearly to most relevant man pages.)

This patch allows calling the Python timezone helper to get the system
timezone before establishing a new PMAPI context and calling the helper
with a pmOptions object to allow figuring out the correct timezone based
on the context time origin (which, after a call to the set_timezone, for
example via the prepare_execute helper, utilizes the underlying OS to
get the correct timezone for the context regardless of DST or such),
thus we get the correct timezone info on Python level without additional
Python modules like pytz. Since the helper returns the timezone as POSIX
offset (instead of, say, something like :Australia/Melbourne) it is easy
to convert this to the more common UTC offsets (for which we had a
helper already).

Comments clarified, code reorganized, QA adjusted, workarounds dropped.