qa/661: handle json output from pmwebd that is not determinstic
The order of groups and the order of details within a group can
both change.

No choice but to syphon off the output to 661.full (to assist triage
if needed), strip the trailing , that is after all elements except
the last in a group, and sort the output ...  and repeat for each
subtest that shares this affliction.