mingw64 build cleanup
Fixes to allow build to run to completion in the Fedora x86_64 mingw64
cross-compilation environment.  Mostly (but not exclusively) in
exec.c for libpcp.

Also fixes to address many of the compilation warnings ... we're down
to less than 50 now!

Lots of these are because the mingw64 headers define getpid() to
return an int, not a pid_t ... leading to (pid_t)getpid() casts
that are no-ops for everyone else.