pmwebd: tolerate temporarily unavailable pmcd for -h permanent contexts
When starting "pmwebd -h $FOO", for establishing a permanent pmapi
context, pmwebd needs to keep the permanent context# available, even
if $FOO's pmcd goes up and down, or indeed if it wasn't initially
running at all.  It now does this by storing the -h $FOO argument in a
webcontext field, and (re)creating the context on demand, if it was
ever broken.  This is a weaker reconnection strategy than pmlogger's,
but good enough in the pmwebd usage context.