src/pmlogger/src/pmlogger.c: cleanup pmcpp properly
Frannk noted pmcpp zombies from pmlogger ... this is not related
to the popen/pclose changes in libpcp nor in pmlogger, but a long
standing (and previously not noticed) bug in pmlogger where pmcpp
was launched on the end of a pipe, but when done, pmlogger called
fclose() not pclose().

qa/577 is new and demonstrates the bug and confirms the fix.