python api: handle Unicode strings with python 2
If Unicode strings are encountered, with Python 3 clients they are
printed as expected but with Python 2 we see nothing as this fails
in pmAtomValue (since the string is not encoded UTF-8 in Python 2):

    c_api.PM_TYPE_STRING : lambda x: str(x.cp.decode('utf-8'))

Below is a patch to fix this. It is perhaps not ideal but should be
safe this late in the release cycle, works with all Python versions,
and, last but not least, does not require changes to clients.