src/pmlogger/src/fetch.c: fix memory leak associated with derived metrics
If a pmlogger configuration file contains the names of derived
metrics (either explicitly or implicitly by PMNS traversal), then
pmlogger has to do some slice-n-dice operations on the pmResult
before writing the archive record.  In this process there was a
(could be significant) memory leak for every record written to
the archive.

This is the major part of the memory leak identified in
https://github.com/performancecopilot/pcp/issues/116