src/libpcp/src/auxserver.c: plug small memory leak in __pmServerSetLocalSocket()
This routine sets a static pointer to either a static or strdup()
result.  If you call it repeatedly and the strdup() path is taken
we leak memory.

Now it always returns a strdup()'d value and takes care to free()
the previous value as needed.

A very small part of the memory leak identified in
https://github.com/performancecopilot/pcp/issues/116