github #350: pmchart(1): Fix memory use after free reported by valgrind.
StatusBar::StatusBar was attempting to replace hte layout of the QStatusBar
from which it inherits. This was done in order to achieve the desired spacing
of the items within the status bar. However, this appears to be not possible
because QStatusBar maintains its own layout structure internally.

This change retains the layout created by QStatusBar and achieves the desired
spacing by traversing the layout and removing all spacing items and margins.