debian build: fix for binary-indep packages
Turns out this obscure error from Makepkgs ...

dh_compress: Can't cd to debian/pcp-conf: No such file or directory

was because the patch we received contained an error and dh_compress
should have been called with a -i at this point.

This changes the heuristics for the files that are selected to
be compressed and so the pcp-doc manifest needed to be updated
(some files that ended in .gz no longer need the .gz because they are
not compressed now).