libpcp: support for help text in PCP archives
Using similar techniques to those we developed for
storing metric labels in archives, we've now also
added recording of metric and indom help text.

Some initial analysis on expected size increases
by Mark and myself suggests this wont become too
onerous, and with Daves compression for metadata
arriving shortly, we expect this will not be too
large an increase in size to be of concern.