python pmcc: add MetricGroupManager.checkMissingMetrics
Adds a new python convenience class (pmcc) method to check if
the argument list of metrics are all available from the current
context. Returns a list of missing metrics, or None if all needed
metrics can be looked up with pmLookupName(3). Calls pmLookupName()
on the entire list first and if (and only if) that fails, then it
checks each individual metric.

Class library fix for Fedora BZ 1508026