pmdaprometheus: permit dots in .url file name
It is sometimes convenient for a pmdaprometheus configuration to use
dot.ted.name.url files.  This way, the PMNS structure for related .url
data sources can be nested siblings.  The man page gives this example:
     FOO.BAR.url
     FOO.BAZ.url
which will generate FOO.BAR.* and FOO.BAZ.* metrics, and cautions against
using this at the same time:
     FOO.url
because of possible PMNS conflicts.