libpcp_pmda: hide pmdaNodeLookup()
This (and a bunch of other methods that will be handled later) method
is private to libpcp_pmda ... the declaration has been removed from
pmda.h and moved to libdefs.h.

Unfortunately, we cannot clawback the symbol's escape from libpcp_pmda
(it has been in exports for ages).  Fortunately there is no man page.

This change is a precursor to hiding the internals of the PMNS data
structures in libpcp.