lots of places: clean up after hiding the PMNS internal data structures
Going forward, if we need libpcp.h to build something, I am going to
ensure all the PCP #includes are of the form
	<hash> include "foo.h"
not
	<hash> include <pcp/foo.h>
to re-inforce the fact that this code is not expected to build outside
the PCP tree.  The <pcp/foo.h> form is still fine for code that can
be built inside or outside the tree (e.g. the source code PMDAs),
but the goal is that these will not need libpcp.h ever, otherwise
the decision to hide stuff in libpcp.h needs to be revisited.