pmmgr: support per-hostid configuration overrides, multiple static-hostids
Add support for configuration items that may be overridden on a
per-hostid basis, to custom-configure individual targets, without
requiring a whole separate pmmgr configuration directory.

Generalize the static-hostid machinery to permit a single pmcd to
present as several distinct hostids.  In conjunction with the above,
this allows multiple concurrent but distinct monitoring configurations
for the same host, for example to segregate different metrics into
different archives, and to manage their retention differently.