libpcp: refactor ... hide more routines
This commit takes the routines that are used only within libpcp but
with prototypes in headers and moves their definitions from libpcp.h
to internal.h.

In a handful of cases, the routines are only used within a single
source file in libpcp and these have become static in that source
file and hidden in the symbols exported from libpcp.

This is an ABI breakage, but in practice no one will notice.