debian packaging: enable systemd
For a long time we've been explicitly disabling systemd functionality
in PCP on Debian-based builds.

As of Ubuntu 17.10 this is no longer an option as the alternative
SystemV mechanisms would appear to have been retired (the sysv-rc-conf
package is not available).  So time to bite the bullet.

Works just fine on Ubuntu 17.10, need to wait on the verdict from
the other QA machines to see if this should remain unconditionally
enabled, or needs some platform-version special sauce.