pmwebd graphite mode: let *._ metric track actual archive boundaries
Previous version of the code that handles the FOO._ pseudo-metric used
the incoming query's time bounds as the interval for the output time
series.  This would mislead to the requester that pmwebd has some data
for the given host over the entire query interval.  Now, pmwebd
intersects the query and the archive bounaries.  qa/1225 enlarged.