pmconfig.py: add transparent support for per-item fetching
In commit ede27b6961 we switched to always use extend_indom
instead of extend_item to resolve RHBZ#1377464 and it helps
with that particular issue (pmrep slow startup time).

But when filtering the results it may be better to retrieve
only the few requested instances, not possibly hundreds of them.

IOW, we want to push instance filtering to C level (libpcp).

We extend pmconfig.py to allow clients to choose either option
transparently; we add a small helper class to provide consistent
interface to retrieve the results regardless of the method
initially chosen. If and only if the current instances only and
also filtering was requested, then instead of extend_indom we use
extend_item. We are careful to add only those instances that
are really present at the time when adding them to pmfg but
also prevent unexpected situation to have a metric w/o instances.

For existing clients this is a transparent change but for clients
to use this we need to add another argument to the validate_metrics
method, we do so in a slightly incompatible manner to make the
parameter names more understandable in the long run. Since the
needed change is trivial and we're doing various API changes in 4.0
anyway, this should be ok as a one-off change.

Follow-up patches will start utilizing this functionality, notable
performance benefits in certain cases included.