pmdaprometheus: rework instance-name prefixing
Former code danced around the issue of how to persist instance names
that are prefixed with instance numbers (because they come from
non-PCP prometheus sources, etc.).  There was a mixture of prefixed
and non-prefixed lookup tables in memory and on disk.

Now, we only store and represent -unprefixed- instance names within
the python pmda, and on disk.  Prefixes get applied, only for those
metric indoms that require it, at the cpmda replace_indom() call.

While in the vicinity, dropped the ' single-quoting of label values in
this same prefixed category.  These were not sufficiently robust to
ensure a unique mapping from prometheus key/value pairs, and were
unsightly to users.  The new scheme has minimum punctuation:

          "number key:value key2:value2"
as in
prometheus.source1.metric1
    inst [0 or "0 abc:0"] value 456
    inst [1 or "1 def:123"] value 123
    inst [3 or "3 ghi:999 mno:828"] value 654
    inst [4 or "4 jkl:998 zoo:kks"] value 236