Add explicit libpcp.h dependencies in makefiles
libpcp.h is now somewhat more fluid than the precursor impl.h, and so
when changes are made to libpcp.h it would be nice if a make across
the src and qa directories would rebuild everything that a libpcp.h
change may have broken.

I've not added dependencies for the other include files as these
are (a) ominpresent, and/or (b) not subject to change at this stage.

This is scaffolding ahead of the pending "fix archive concurrency"
and __pmArchCtl/__pmLogCtl changes.

And yes I know this sort of dependency in the makefiles is brittle
and subject to change, but that's the lesser of two evils.