libpcp: remove __pmLogCtl from internal routines
I am changing to using __pmArchCtl in lieu of __pmLogCtl in all
the internal routines involving manipulation of PCP archives.

This is because some fields in __pmLogCtl will be migrating to
__pmArchCtl, and we can always use ac_log to navigate from
__pmArchCtl to __pmLogCtl, but the reverse navigation is not
possible.

For some of the utilities that had a private and fake instances
of __pmLogCtl, they now need private and fake instances of both
__pmArchCtl and __pmLogCtl, with ac_log set up correctly.