pmdaprometheus: add labels callbacks
Add support for labels back into the Prometheus PMDA.

Note there is one "end-point source" at the pmid cluster level,
which has a label for the end-point URL ("url" label) and its
hostname ("hostname" label).

file:// URLs use the local host name, as is returned by
socket.gethostname(). Any http or https URLs for "localhost" are
also labeled as the local host name (rather than "localhost").
Remote URLs extract the host name from the URL. The "source"
label at the cluster level is also extracted from the URL, and
is the same as the second level of the dynamic pmns entry for
the end-point URL.

Other labels include the "instance" label at the PMID item level,
which is the JSONb formatted labelset taken from the fetched
Orometheus labelset. This is similar to the PCP instance name, but
without the unique instance ID prefix.