pmdasimple: make more compelling example use of labels
Instead of setting labels that are simply existing metric
metadata (cluster/indom/... identifiers, etc) switch over
to more "interesting" labels like units/unitsystem for the
sec/min/hour case, and an indom_name label for both indoms.

Drop the cluster labels since they're not adding value and
the "time" one is even incorrect (control metrics are also
in that cluster).  Add a comment that its not necessary to
provide callbacks for all types of label identifier - using
the default handler is fine in these situations.

Since indoms cannot actually be named (no PMNS for indoms),
currently, it may actually be a good idea to go through and
add a human readable indom_name label for all of the indoms
(in other PMDAs too, IOW).