build: generalise the derived metrics makefile
Change the build for derived metrics configurations so
that the makefile does not need to be updated for new
additions.  Also, always install these files, not just
for Linux - this way, tools running on other platforms
can still monitor archives or remote Linux hosts using
the set of dervied metrics defined globally.