pmrep/pcp2xxx: introduce per-metric precision support
We extend the compact per-metric specification as follows

-metric[,label[,instances[,unit/scale[,type[,width]]]]]
+metric[,label[,instances[,unit/scale[,type[,width[,precision]]]]]]

in a backward-compatible and transparent manner for existing configs.

With this we can later on get rid of some offending decimals when
emulating some of the well-known tools. All existing QA ok (as it
should be), add few quick tests to cover this area as well.