pmstore: add -F command line option
Normally pmstore reports the "new" value from the command line.
But for some metrics the "store" operation has different semantics,
e.g. to increment the current value, a trigger that does something
but does not change the value of the metric, etc. and in these cases
a pmFetch after the pmStore returns a more accurate "new" value than
the value from the command line ... -F does just this.