pmlogextract: change <mark> strategy
If prologue and epilogue records are available in the input archives
use these to determine if pmcd's state is unchanged between the
end of one archive and the next output record from another archive
(typically the first record from the next archive in temporal order
for the log merging case).

And if the state is unchanged (same pid and same (config) seqnum for
pmcd), then do not emit a <mark> record.

This should remove the <mark> records from daily archive merging for
most stable production environments.

The (new) -m option restores the original <mark> strategy.